Add "popup" invoking attribute, and clean up ancestor walk
With this CL, the "popup" attribute can be used on <button> elements
only, to invoke a popup. Additional elements (e.g. <input type=button>,
other input types, etc.) will be added in subsequent CLs. In addition,
this CL adds storage of all invoking elements on each <popup>, so that
invokers can be checked in the ancestral popup walk. No code has been
added yet to populate this vector, since it is still up in the air [1]
whether this should be part of the behavior, and it is a bit
complicated to maintain this linkage.

[1] https://github.com/MicrosoftEdge/MSEdgeExplainers/issues/463

Bug: 1168738
Change-Id: I62af6e7655ad67540082b5a796f8390e4a890f72