Add type check in CSSScale and implement generate expression node for invert and negate
For the current implementation, the CSS Typed OM could occur null pointer access when building CSSMathExpressionNode. Because in the middle of ToCalcExpressionNode() recursion, nullptr is accessed instead of propagation.

Also, The CSSScale's constructor didn't check the type properly for the expression. This possibly might be needed another typed om.

Therefore, implement TODOs for building CSSMathExpressionNode tree and add condition in IsValidScaleCoord() to check resolved type is number for CSSMathExpressionNode tree.

Bug: 1188610
Change-Id: If6bf70b791b72c7ed67465f898130e098f68e853